requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
backcall==0.2.0
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
decorator==5.2.1
distlib==0.3.9
executing==2.2.0
Faker==35.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2024.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.2.3
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.0.6
WTForms==3.1.2
zipp==3.20.2
